LayoutTests/fast/filesystem/file-metadata-after-write-expected.txt

This verifies File.size (for a file from FileSystem API) always returns the fresh size even after the file is modified.
Writing 1234567890 to the file...
Created a writer.
Write succeeded.
PASS testFile.size is testText1.length
Writing abcdefghijklmnopqrstuvwxyz to the file...
Created a writer.
Write succeeded.
PASS testFile.size is testText2.length
